pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "face-generation-gan"
version = "0.1.0"
description = "WGAN-GP for generating 64x64 celebrity faces (portfolio project)."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "torch",
  "torchvision",
  "numpy==1.26.4",
  "matplotlib==3.8.4",
  "pillow==10.2.0",
  "tqdm==4.66.4",
]

[project.scripts]
face-gan-train = "face_gan.train:main"
face-gan-generate = "face_gan.generate:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I", "B"]
ignore = ["E501"]

